pyproject.toml
# LICENSE HEADER

[tool.poetry]
name = "app"
version = "0.1"
description = "Demo MS"
authors = ["Kuldeep K"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.103.2"
orjson = "^3.9.1"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
uvicorn = "^0.23.2"
black = "^23.3.0"
pytest = "^7.3.2"
unittest2 = "^1.1.0"
loguru = "^0.7.0"
tenacity = "^8.2.2"
pre-commit = "^3.3.3"
